Create Python programming for character encryption
I watched a movie a while ago called the Imitation Game based on Alan Turing decrypting German cyphers during WW2 and I was tempted to try it out for myself. 

During this, I learnt how to: 
- format using f strings 
- use functions to retrieve alphabets and characters instead of writing them all out 
-Ito duplicate a variable into another variable 
- using index to assign the same letter to it's encrypted counter part 

Seeing these processes from a distance makes it look so complex but once I tried it for myself, I found it to be quite doable. Makes me wonder what else is within grasp for me to learn that I never thought possible