Create Hash generator
I was learning about blockchain and wondered if I could create some of my own hashes and test the mechanisms and this was very fun to learn. A bit heavy on the syntax as a beginner to python but eventually, I got the hang of it.

I learned how to use a library with hashes stored and print it out from their stored mode. I learned how to format a list and print the items in the list. 

Needless to say this was a very informative exercise and I am very grateful to have been guided by the following video:
https://youtu.be/SVmZZK11Tjc?si=JKehd2tJxvkrgCRC
It was a wonderful way to build my knowledge base